requirements.txt
flask
opencv-python
torch
torchvision
joblib
scikit-learn
numpy
pandas
pillow
psutil
